feat: create StudentModulesPageClient component with course data and layout
